+++ AgriculturalWarehouseManagement/src/main/java/com/example/AgriculturalWarehouseManagement/Backend/repositorys/blog/BlogCategoryRepository.java
@@ -3,5 +3,6 @@ package com.example.AgriculturalWarehouseManagement.Backend.repositorys.blog;
 import com.example.AgriculturalWarehouseManagement.Backend.models.BlogCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogCategoryRepository extends JpaRepository<BlogCategory, Long> {
+public interface BlogCategoryRepository extends JpaRepository<BlogCategory, Integer> {
+
 }
